server/core-service/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
backoff==2.2.1
certifi==2024.8.30
click==8.1.7
coloredlogs==15.0.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
flatbuffers==24.3.25
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
numpy==2.1.3
ocrmac==1.0.0; platform_system == "Darwin"
onnxruntime==1.20.1
opencv-python==4.10.0.84
openvino==2024.5.0
openvino-telemetry==2024.5.0
packaging==24.2
pillow==11.0.0
protobuf==5.29.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
pyobjc-core==10.3.1; platform_system == "Darwin"
pyobjc-framework-Cocoa==10.3.1; platform_system == "Darwin"
pyobjc-framework-CoreML==10.3.1; platform_system == "Darwin"
pyobjc-framework-Quartz==10.3.1; platform_system == "Darwin"
pyobjc-framework-Vision==10.3.1; platform_system == "Darwin"
python-bcrypt==0.3.2
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
rapidocr-onnxruntime==1.4.0; platform_system != "Darwin"
rapidocr-openvino==1.3.25; platform_system != "Darwin"
rich==13.9.4
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.13.3
tqdm==4.67.1
typer==0.13.1
typing_extensions==4.12.2
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
